Refactor `ApInt` and `ApsInt` constructors: remove unused `sizeInBytes` parameter and adjust related logic in `SemaHelpers`.

+++ src/Sema/Constant/ConstantValue.cpp
@@ -333,7 +333,7 @@ ApsInt ConstantValue::getIntLike() const
     if (isRune())
         return ApsInt(getRune(), 32, true);
     if (isNull())
-        return ApsInt(0, 64, true);
+        return ApsInt(static_cast<uint64_t>(0), 64, true);
     SWC_UNREACHABLE();
 }
